pyproject.toml
[project]
name = "watt-monitor"
version = "0.1.0"
description = "Monitors AC power consumption via CT sensor on Raspberry Pi and sends metrics to InfluxDB"
requires-python = ">=3.11"
dependencies = [
    "gpiozero>=2.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rpi-gpio>=0.7.1 ; sys_platform == 'linux'",
]

[dependency-groups]
dev = []
